refactor(mcp-server): split economy.go and refdata.go by domain
Split economy.go (955 lines) into:
- economy.go: shared interfaces, types, register, helpers (136 lines)
- economy_manifest.go: validate/plan/apply tools (413 lines)
- economy_graph.go: graph/history/rollback tools (426 lines)

Split refdata.go (873 lines) into:
- refdata.go: interfaces, register, structure tool, helpers (216 lines)
- refdata_instrument.go: instrument list/describe tools (193 lines)
- refdata_saga.go: saga/handler/market data tools (482 lines)

Remove both from knownOversizedFiles - all files now under 800 lines.
