+++ tests/architecture/size_test.go
@@ -28,8 +28,6 @@ var knownOversizedFiles = map[string]bool{
 	"services/current-account/service/saga_handlers.go":          true,
 	"services/identity/service/server.go":                        true,
 	"services/market-information/service/observation_service.go": true,
-	"services/mcp-server/internal/tools/economy.go":              true,
-	"services/mcp-server/internal/tools/refdata.go":              true,
 	"services/party/adapters/persistence/repository.go":          true,
 	"shared/pkg/saga/schema/schema.go":                           true,
 }
